pyproject.toml
[project]
name = "prosaia"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel",email = "carloscgmorais@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]

[tool.poetry]
packages = [{include = "prosaia", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
ruff = "^0.12.5"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]


[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]


[tool.ruff.format]
preview = true
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"
run = "fastapi dev src/prosaia/main.py"
pre_test = "task lint"
test = "pytest -s -x --cov=src/ -vv"
post_test = "coverage html"
